base_cff is now only containing the postfixes
